fix py3 number of rows

+++ spectrogram.py
@@ -137,7 +137,7 @@ class Spectrogram(Snuffling):
 
         p = None
 
-        ncols = len(nslcs) / 5 + 1
+        ncols = int(len(nslcs) / 5 + 1)
         nrows = (len(nslcs)-1) / ncols + 1
 
         tmin = min(times)
